Fixed JWT race condition
The setting/updating of the JWT was in `componentDidUpdate` which could 
trigger multiple times before the initial request for a JWT update could 
complete. This would lead to the reading of a partial or corrupt config. 
I've limited the JWT updates to only occur when first loading the App, 
and the action is now chained so other requests won't go out until the 
JWT update has completed.