fix: Nested files deleted after one file is renamed
There are times when a User may choose to only rename one file in folder of files. After the rename, the folder and all of it's child files were being deleted. I previously had a safeguard against that, but the comment I left for that safeguard wasn't clear.
I've reimplemented the old safeguard, alongside the previous patch for the Unhandled rejection.