feat: On assignment of new id, preview all that match
Previously, after a User assigned a new id to an item - and there were other items of the same name in the list - only one item would be previewed.
Now any items in the list with the same `lookupName` will also get previewed, saving the User from having to reload the App to get the changes.