fix: Anime episodes not found
Turns out there are some Anime series which use the total number of series episodes for the `episode_number`. To account for that, I just use the season's episode index (which seems to be accurate for now), and I prepend the listed episode number to the episode name to provide as much info as possible.