fix: statusCode error and inefficient series lookups
When looking up a new series with a lot of episodes, a stacktrace was being thrown on the Server because `resp` was `undefined` during a series lookup. That fix was simple, but led me to find I was making a series & episodes request for all items if the series wasn't cached. So, if there were 100 episodes that needed renaming, that was 100 requests for the same series and it's episodes. I was basically DDOS'ing theMDB API.
Now the Server's smart enough to scrape the new series, then do the lookups for the names.