fix: Assign errors after name update in GUI
There's an edge-case where a User could change a name in the GUI, which would generate a new `lookupName`, but then click Preview which would then try to generate a `lookupName` against the original file name which will return `null` in some cases. Added logic to utilize the previously generated `lookupName` only if a new one hasn't been generated.