fix: Assigned item removing year after cache update
If a User previously assigned a series that had a year in it (like there were multiple series matches, and they chose a newer one so the year was added), and then clicked the update cache button - a new cache would be generated without the year, and all series names will now be without a year and a new folder would be created.

Pretty particular edge case, but it was solved by always keying off of the cached `cacheKey` and specifying that certain flows should execute if an update was requested.