feat: Account for misc. encoded characters
`xmlbuilder2` accounts for and encodes most odd characters to html entities, but there were some outlier characters (like magic quotes) that it misses. Normally I'd use `he`, but there's no way to intercept the `text` node creation, which led to double-encoding. There's a `parser: text` option, but there doesn't seem to be a way to tell it not to encode, so I manually sanitize text entries where User's are most likely to have copied and pasted bad characters.