fix: Odd visual behavior in Replace dialog
If a name had something like `Name (2019) - 1x01` and a pattern like `(\d{1})x(\d{2})`, the markup was wrapping the first group correctly, but then wrapping the second group value (`01`) in the year `(2<mark>01</mark>9)`.

The fix was simply utilizing the same match index that the RegEx uses. Once I limited the wrapping logic to only effect the text at the index where the matches started being found, things started behaving as expected.