feat: Filter fanart.tv images by language
The scraper uses the highest rated image for each category. A majority of the time, those images have English text (where applicable), but sometimes the highest rated item will be in another language.
So I added a filter to all images to only accept items with English text, or no text at all.