feat: Ability to pad numbers in Replace
If a User was using the index for episodes, and they were iterating say a range of files from 1-10, they'd end up with something like `s01e1` and `s01e10`. They can account for the items below 10 manually in the GUI or with another replace, but why?
Users can now have something like `[#+00]` which will translate to `s01e01`.