feat: Prevent accidental Replace
Since the input values in Replace are persisted, it's possible that a User could accidentally try replace a name with one that already exists which would cause the old file to be overwritten by the newly renamed item - which could be really bad if they're changing a list of files.
Now a list of errors will be returned informing the User which files weren't renamed.