fix: cache lookups sometimes fail
Since there are series that end with a year-like format (like "Mystery Science Theater 3000"), series names have to have years wrapped in parenthesis like "name (year)" to ensure proper lookups.

While working on test updates I came across a race-condition where the request for the names was going out before the `ids` were loaded. This meant that a cache name was being generated on the Server and a lookup was occurring. The lookup would always fail because the name being sent to the Server was "name year", but when the mapped `ids` were present, it was being mapped to "name (year)" since it was previously assigned.

This fix ensures that if a year is found in a name, it'll be sent to the Server in the proper format.