ops: Separated tests
Running and writing tests was a bit of a pain because each case was dependant on the previous stage. Mainly because I was trying to prevent the extra load/prep time of a fresh page. Turns out that it doesn't take long at all to re-init a page with the generated data. This new setup actually shined some light on existing issues as well.

- Added some logic to `conf.app` so that it could be loaded in the tests without error.
- Added in `resetData` to clear out screenshots and revert any changes made to `cache` files after test runs. This meant I had to change how folders and files were mapped in the test container (added in a parent `/repo` path) which allowed me to run `git` commands as I would outside of a container.
- Refactored all tests so that they could be run individually.
- Added in logging within test setup steps to verify that things are running in the expected order.