fix: Old files not being overwritten
I previously put in a fix to not overwrite an existing file while running a Replace.
That logic was a bit greedy, and was running during the Rename as well, which meant that if there was a file of the same name in the destination, the new file would no longer overwrite it (which it used to do).